requirements.txt
FastAPI
uvicorn[standard]
pydantic-settings
sqlalchemy
asyncpg
python-dotenv